TAF-14 Moving Config parts to test-common; added dependency for skyxplore- to test-common
