+++ skyxplore-ui-test/build.gradle
@@ -122,6 +122,7 @@ task smokeTestAndReport(type: GradleBuild) {
 }
 
 dependencies {
+    implementation project(':test-common')
     testCompileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
     compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.26'
     runtime group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.11.2'
