+++ skyxplore-ui-test/src/test/java/com/mentoring/framework/Config.java
@@ -2,11 +2,9 @@ package com.mentoring.framework;
 
 import com.mentoring.framework.utils.EmailUtils;
 import com.mentoring.model.Browser;
+import com.mentoring.utilities.ConfigUtils;
 import lombok.extern.slf4j.Slf4j;
 
-import java.io.IOException;
-import java.util.Arrays;
-import java.util.Objects;
 import java.util.Properties;
 
 @Slf4j
@@ -18,11 +16,9 @@ public final class Config {
     private static final String SKY_XPLORE_LOCALHOST = "skyxplore-localhost";
     private static final String SKY_XPLORE_PROD = "skyxplore-prod";
     private static final String PATH_WEB_EXTENSION = "/web";
-
     private static final String BASE_URL;
     private static final String PORT;
     private static final String ENVIRONMENT = "ENVIRONMENT";
-    private static String[] validEnvironmentConfigurationNames = {SKY_XPLORE_LOCALHOST, SKY_XPLORE_PROD};
     private static final String DATABASE_NAME;
     private static final String JDBC;
     private static final String DB_USER;
@@ -30,7 +26,7 @@ public final class Config {
 
 
     static {
-        Properties prop = loadProperties();
+        Properties prop = ConfigUtils.loadProperties();
         LOAD_WAIT = Integer.parseInt(prop.getProperty("load_wait"));
         BROWSER = Browser.parse(System.getenv("BROWSER"));
         PORT = prop.getProperty("port");
@@ -52,45 +48,6 @@ public final class Config {
     }
 
     private Config() {
-    }
-
-    private static Properties loadProperties() {
-        String testEnvironmentName = System.getenv(ENVIRONMENT).toLowerCase();
-        String propertiesFileName = validateEnvironment(testEnvironmentName) + "-config.properties";
-        return loadProperties(propertiesFileName);
-    }
-
-    private static Properties loadProperties(String propertyFile) {
-        Properties properties = new Properties();
-        try {
-            properties.load(Objects.requireNonNull(Thread.currentThread().getContextClassLoader().getResourceAsStream(propertyFile)));
-        } catch (IOException e) {
-            log.error("Error loading the " + propertyFile + " file", e);
-        }
-        return properties;
-    }
-
-    private static String validateEnvironment(String currentEnvironment) {
-        if (currentEnvironment == null) {
-            log.error("Environment is not set. Please choose from: {}", Arrays.toString(validEnvironmentConfigurationNames));
-            System.exit(1);
-        }
-        return compareCurrentEnvironmentWithValidEnvironments(currentEnvironment);
-    }
-
-    private static String compareCurrentEnvironmentWithValidEnvironments(String currentEnvironment) {
-        String validEnvironmentName = "";
-        for (int i = 0; i < validEnvironmentConfigurationNames.length; i++) {
-            if (validEnvironmentConfigurationNames[i].equalsIgnoreCase(currentEnvironment)) {
-                validEnvironmentName = validEnvironmentConfigurationNames[i];
-                break;
-            } else if (i == validEnvironmentConfigurationNames.length - 1) {
-                log.error("Failed loading environment configuration. Given environment name '{}' is not valid."
-                        + " Please choose from: {}", currentEnvironment, Arrays.toString(validEnvironmentConfigurationNames));
-                System.exit(1);
-            }
-        }
-        return validEnvironmentName;
     }
 
     public static boolean isLocalEnvironmentUsed() {
