+++ test-common/build.gradle
@@ -1,16 +1,42 @@
-plugins {
-    id 'java'
-}
-
 group 'com.mentoring.ta'
 version '1.0-SNAPSHOT'
 
-sourceCompatibility = 1.8
+buildscript {
+
+    repositories {
+        jcenter()
+        mavenCentral()
+    }
+    dependencies {
+        classpath "io.qameta.allure:allure-gradle:2.7.0"
+    }
+}
+apply plugin: 'java'
+
+compileJava.options.encoding = "UTF-8"
+compileTestJava.options.encoding = "UTF-8"
+
+repositories {
+    jcenter()
+    mavenCentral()
+}
 
 repositories {
     mavenCentral()
 }
 
 dependencies {
-    testCompile group: 'junit', name: 'junit', version: '4.12'
+    compile group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
+    compile group: 'org.slf4j', name: 'slf4j-api', version: '1.7.26'
+    runtime group: 'org.apache.logging.log4j', name: 'log4j-core', version: '2.11.2'
+    compile group: 'org.slf4j', name: 'slf4j-log4j12', version: '1.7.26'
+    runtime group: 'ch.qos.logback', name: 'logback-classic', version: '1.2.3'
+    runtime group: 'ch.qos.logback', name: 'logback-core', version: '1.2.3'
+    compile 'net.javacrumbs.json-unit:json-unit:1.28.1'
+    compile group: 'javax.mail', name: 'mail', version: '1.4.7'
+    compile group: 'org.jsoup', name: 'jsoup', version: '1.11.3'
+    compile group: 'javax.activation', name: 'activation', version: '1.1'
+    compile 'io.rest-assured:rest-assured:3.3.0'
+    compile group: 'mysql', name: 'mysql-connector-java', version: '8.0.16'
+    testCompile group: 'org.awaitility', name: 'awaitility', version: '4.0.0-rc1'
 }
